Fixed faulty file hashes

+++ game/shared/ms/crc/crchash.h
@@ -1,7 +1,7 @@
 #ifndef MSSHARED_CRCHASH_SHARED_H
 #define MSSHARED_CRCHASH_SHARED_H
 
-bool MatchFileCheckSum(const char* FilePath, unsigned long CheckSum);
-unsigned long GetFileCheckSum(const char* FilePath);
+bool MatchFileCheckSum(const char* FilePath, unsigned int CheckSum);
+unsigned int GetFileCheckSum(const char* FilePath);
 
 #endif
